Refactor collection management by removing the Indexer trait and related files, enhancing request indexing functionality
