+++ crates/moss-collection/src/lib.rs
@@ -1,12 +1,9 @@
 pub mod models;
 
 pub mod collection;
-pub mod indexing;
 pub mod kdl;
 pub mod storage;
 
-// pub mod collection_handle;
-// pub mod collection_manager;
 pub mod request_handle;
 
 pub mod leased_slotmap;
